Graph inclusion
Here I've moved in the jcommon/jfreechart libraries along with some work
for the graph class.

I've been having trouble with the graphs in Maven so I made a java file
called first where i've been building the data for testing as well as
experimenting with the chart types.

In the first file is all the pie charts, half the bar charts, and 1 of
the 4 line charts that will be included.

Next I will be moving all datasets into the main func and passing the
data as a parameter. Then I will be working on getting it to work in
Maven (there's some class pathing issue that JPA doesn't like). After
that I'll work on doing the chart image export.

After that I will work on individualizing the graphs, changing colors,
details and stuff.

EDIT: Cannot include JCommon & JFreeChart Libraries here, will do
seperate commit